requirements.txt
# 📁 최종 배포용: backend/requirements.txt
# FastAPI 라이브러리 관리 시스템 - 백엔드 의존성

# FastAPI 및 웹 서버
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# 데이터베이스
sqlalchemy==2.0.23
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1

# AWS S3
boto3==1.34.0
botocore==1.34.0

# 인증 및 보안
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-decouple==3.8

# 유틸리티
pydantic==2.5.0
pydantic-settings==2.1.0

# 개발 도구
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0

# 기존 의존성들 (호환성 유지)
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
click==8.3.1
colorama==0.4.6
greenlet==3.3.0
h11==0.16.0
idna==3.11
pydantic_core==2.41.5
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0